requirements.txt
torch>=1.7.0
torchvision>=0.8.0
numpy>=1.19.0
opencv-python>=4.5.0
gymnasium>=0.29.0
stable-baselines3>=2.2.0
matplotlib>=3.3.0
seaborn>=0.11.0
pyyaml>=5.4.0
transformers>=4.36.0